Fix up unsatisfied links by cleaning up definitions Add new compiler flags support for nd4j-native Add validation for missing method definitions More clean up on ints/unsigned values to sd::Longs

+++ libnd4j/include/ops/declarable/helpers/cpu/indexReductions.hpp
@@ -120,7 +120,7 @@ static SD_INLINE void indexInnerReductionConstRank(const X* buffer, X& current, 
 }
 
 template <typename X, typename Z, typename ReductionOp, bool LastIndexFaster = true>
-static SD_INLINE void indexInnerReduction(const int& rank, const X* buffer, X& current, Z& argCurrent,
+static SD_INLINE void indexInnerReduction(const LongType& rank, const X* buffer, X& current, Z& argCurrent,
                                           const sd::LongType* bases, const sd::LongType* strides,
                                           const sd::LongType& outerLoopStart, const sd::LongType& outerLoopStop,
                                           const sd::LongType& innerLoopCount) {
